Serialization long type incorrect for Linux 32-bit, Add more tests for ReliableTopic for increasing coverage (#155)
* Added more tests for ReliableTopic.

* A fix is done at ReliableTopicExecutor.

* Linux portability fix.

* Changed long to int64_t for serialization.

+++ hazelcast/src/hazelcast/client/serialization/ObjectDataInput.cpp
@@ -63,8 +63,8 @@ namespace hazelcast {
                 return dataInput.readInt();
             }
 
-            long ObjectDataInput::readLong() {
-                return (long)dataInput.readLong();
+            int64_t ObjectDataInput::readLong() {
+                return dataInput.readLong();
             }
 
             float ObjectDataInput::readFloat() {
